Change int to long

+++ features/merchant/review/src/main/java/com/tokopedia/review/feature/historydetails/presentation/viewmodel/ReviewDetailViewModel.kt
@@ -68,7 +68,7 @@ class ReviewDetailViewModel @Inject constructor(private val coroutineDispatcherP
         return userSession.userId
     }
 
-    fun submitReputation(reputationId: Int, reputationScore: Int) {
+    fun submitReputation(reputationId: Long, reputationScore: Int) {
         _submitReputationResult.value = LoadingView()
         launchCatchError(block = {
             val response = withContext(coroutineDispatcherProvider.io) {
